requirements.txt
playwright==1.41.1
pytest==8.3.5
pytest-playwright==0.7.0
pytest-html==4.1.1
python-dotenv

